Add a common processor to parse currencies.
This commit contains the processor and associated tests. A new helper function called `process` was added in the utilities module to loop over all resources and rows. This is an attempt to standardize the way processors are written. The helper function also has tests. I suggest to include this helper directly in the `wrapper` module of the pipeline framework.