alternative fiscal schema
I tried to make a work around for the activity/recipient information as these need to be linked. I gave those two a functional classification, and the fund information the administrative classification. Also Geo-code and hierarchies proved to be difficult to map, again I tried a work-around, but it's not pretty.